update itemTypes.basePrice to current live prices in Jita market (thanks Groove) update all item basePrice variables to double change itemTypes.basePrice to double add market config options to config xml updated version info updated market seeding sql to new price updates updated Groove's item price script to output sql

+++ src/eve-server/corporation/CorpStationMgr.cpp
@@ -250,7 +250,7 @@ PyResult CorpStationMgrIMBound::Handle_SetCloneTypeID(PyCallArgs &call) {
     }
 
     //Get cost of clone
-    int cost = m_db.GetCloneTypeCostByID(arg.CloneTypeID);
+    double cost = m_db.GetCloneTypeCostByID(arg.CloneTypeID);
 
     //take the money, send wallet blink event record the transaction in their journal.
     std::string reason = "DESC: Updating Clone in ";
